lambda/requirements.txt
PyJWT==2.8.0
bcrypt==4.0.1
boto3>=1.26.0
